Implemented the image deletion function for when a user is removed.

+++ Workshops/Workshop 7/users/delete.php
@@ -1,17 +1,20 @@
 <?php
-include_once('../inc/validateSession.php');
 if ($_GET) {
     include_once('../utils/database.php');
     include_once('../utils/functions.php');
     $id = $_GET['id'];
 
-    $database = new databaseManager();
+    if ($id) {
+        $database = new databaseManager();
+        $relativePathImage = $database->getUserPathImage($id);
 
-    if ($database->deleteUser($id)) {
-        $database->closeConnection();
-        header("Location: /Workshop%207/users/users.php");
-    } else {
-        header("Location: /Workshop%207/users/users.php?error=true");
+        if ($database->deleteUser($id)) {
+            $database->closeConnection();
+            deleteImage($relativePathImage);
+            header("Location: /Workshop%207/users/users.php");
+        } else {
+            header("Location: /Workshop%207/users/users.php?error=true");
+        }
     }
 } else {
     header("Location: /Workshop%207/users/users.php");
